refactor(sdk): split types/index.ts into focused contract modules
The 1,524-line SDK types file is split into six cohesive modules behind
the same barrel: primitives, manifest, runtime, services, registration,
context. index.ts stays the @makinbakin/sdk/types entrypoint
(scripts/build-sdk-package.ts SDK_EXPORTS + the vendor bundle) and
re-exports the full surface via `export *`. Cross-module references use
type-only sibling imports; no module reaches @bakin/core or repo source,
so the published self-containment holds. The 8 dead misc types
(Calendar*, Memory*, ProjectMeta, the SDK's orphaned Heartbeat and
WorkflowStep — zero importers) are dropped in the move. Exported surface
is otherwise identical (125 → 117 = the 8 deletions); pure types, zero
runtime. Verified: typecheck, full suite, build:vendors, name-set diff.
